Code/requirements.txt
Flask==1.1.2
numpy
onnxruntime
opencv_contrib_python==4.5.5.64
opencv_python==4.7.0.72
Pillow==9.5.0
pycocotools==2.0.6
torch==1.9.0
torchvision==0.10.0
tqdm==4.50.2

